# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ProjectsCreateDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"start_date": {
						"type": "string"
					},
					"end_date": {
						"type": "string"
					},
					"cost": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"start_date",
					"end_date",
					"cost"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ProjectsCreateDTO-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ProjectsCreateDTO"
						},
						"type": "array"
					},
					"error": {},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"error": {},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectsEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"start_date": {
						"type": "string"
					},
					"end_date": {
						"type": "string"
					},
					"cost": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"start_date",
					"end_date",
					"cost"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HttpStatus": {
				"enum": [
					200,
					201,
					204,
					400,
					401,
					403,
					404,
					409,
					500
				],
				"type": "number"
			},
			"IResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProjectsDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"start_date": {
						"type": "string"
					},
					"end_date": {
						"type": "string"
					},
					"cost": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"start_date",
					"end_date",
					"cost"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"error": {},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"HolidaysDTO": {
				"properties": {
					"name": {
						"type": "string",
						"example": "good friday"
					},
					"date": {
						"type": "string",
						"example": "DD/MM"
					},
					"year": {
						"type": "string",
						"example": "YYYY"
					}
				},
				"required": [
					"name",
					"date",
					"year"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_HolidaysDTO-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/HolidaysDTO"
						},
						"type": "array"
					},
					"error": {},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HolidaysEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"year": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"date",
					"endDate",
					"year"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateHolidaysDTO": {
				"properties": {
					"name": {
						"type": "string",
						"example": "good friday"
					},
					"date": {
						"type": "string",
						"example": "DD/MM"
					},
					"year": {
						"type": "string",
						"example": "YYYY"
					},
					"id": {
						"type": "number",
						"format": "double",
						"example": 1
					}
				},
				"required": [
					"name",
					"date",
					"year"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "employee_attendance",
		"version": "0.1.0",
		"contact": {}
	},
	"paths": {
		"/projects": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProjectsCreateDTO-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"projects"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Save",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"statusCode": {
											"$ref": "#/components/schemas/HttpStatus"
										},
										"data": {
											"$ref": "#/components/schemas/ProjectsEntity"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"statusCode",
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"projects"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectsCreateDTO"
							}
						}
					}
				}
			}
		},
		"/projects/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProjectsDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/holidays": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_HolidaysDTO-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Holidays"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Save",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"allOf": [
														{
															"$ref": "#/components/schemas/HolidaysDTO"
														},
														{
															"$ref": "#/components/schemas/HolidaysEntity"
														}
													]
												},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {},
												"message": {},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"error",
												"message",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Holidays"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/HolidaysDTO"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"message",
												"success"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"data": {
													"$ref": "#/components/schemas/HolidaysEntity"
												},
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"data",
												"message",
												"success"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {},
												"message": {},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"error",
												"message",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Holidays"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateHolidaysDTO"
							}
						}
					}
				}
			}
		},
		"/holidays/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Holidays"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Holidays"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}